zoom site map

+++ BuildersAccessPo/Imagezoom/MRZoomScrollView.m
@@ -90,8 +90,8 @@
     return imageView;
 }
 
-- (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(float)scale
-{
+-(void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(CGFloat)scale{
+
     [scrollView setZoomScale:scale animated:NO];
 }
